requirements.txt
# Core data processing
pandas>=2.0.0
pyarrow>=14.0.0
polars>=0.20.0

# API and HTTP requests
requests>=2.31.0
httpx>=0.27.0

# REST API (for api_server.py)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6

# Geospatial analysis
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.15.0

# Database and query engine
duckdb>=0.10.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard/Web apps
streamlit>=1.30.0

# Core dependencies
requests>=2.31.0
pandas>=2.0.0
pyarrow>=12.0.0

# Data validation
pydantic>=2.5.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
ipykernel>=6.27.0
jupyter>=1.0.0
black>=23.12.0
pytest>=7.4.0

# Code formatting
black>=23.0.0
isort>=5.12.0
